Removed compile error in VS8 on line 256. The 'operator <<' on that line is ambiguous because it could be:  operator <<(double)  operator <<(float)  operator <<(unsigned long)  operator <<(long)  operator <<(unsigned int)  operator <<(int)  operator <<(unsigned short)  operator <<(short)  operator <<(unsigned char)  operator <<(char)  operator <<(bool)
Please note that on a 64 bit system size_t is a 64 bit value. Line 256 is trying to match the argument list '(Pieces::DataStream, unsigned __int64)'.

I am casting size_t to an unsigned int now, but I am not sure if this is the correct way to do it.

+++ src/pieces/data_stream.cpp
@@ -253,7 +253,7 @@ DataStream& operator>>(DataStream& ds, ByteArray& ba)
 
 DataStream& operator<<(DataStream& ds, const std::string& str)
 {
-    ds << str.size();
+    ds << static_cast<unsigned int>(str.size());
     ds.writeBytes(ByteArray(str.data(), str.size()));
 
     return ds;
